Compiler: fixed accessing non-static members of static members
It seems like extract_variable_name() is gathering a list of symbols accessed in a chain (i.e. "x.y.z.a"). The problem was that when the first member is a type, it sets up a flag saying that the next member must be static, but does not reset it when next member is a static member which may have non-static members on its own.
This fix resets the flag in that case, allowing the function to continue parsing this chain.