Revert "Timing: use c++ std timers instead of maintaining own timer thread."
This reverts commit f2ab428a596c67d950f2aed5fbd335f6ccc72e3f.

since the commit did a dozen things at the same time, it was very hard to
untangle things that changed since.

the commit doesn't compile yet, since the changes of b78dc01e70c5ed74166ffb95921ec9d085e56561
need to be undone still.

it is possible that some of the many changes to voice audio need to revised
by someone familiar with the code, as there have been many changes that conflicte
and caused merge conflicts.