Engine: fix missed walk-behinds update for 3D renderer
This fixes a rare case when the room's background was drawn upon using raw drawing script functions after room is loaded but before it is shown on screen for the first time.
In such case the first render frame does not have the walk-behinds reupdated, because it does not test for `current_background_is_dirty` flag, only `walkBehindsCachedForBgNum` (which already matches proper background index). This is fixed at the second render frame.

It seems that his mistake was previously ignored because the very first render frame occured at the time when fade-in did not start yet and screen was blacked out and it's the second frame which was actually seen by players. But after commit fb7d8abe8c55cbf0a89bf9aba0c170731aded6f4 engine no longer makes full render pass while the screen is fully fade out, making this issue apparent.