Engine: don't pessimize general execution speed for rare special case
the special case for a push instruction followed by a pop in the code
here gives a penalty in form of an additional branch which is likely
mispredicted,
and several added instruction to every single push call for a relatively
rare special case.
it's better to improve the codegen not to emit such inefficient
statements.