Engine: fixed inconsistent crossfade/dissolve behavior with screen tint
Non-instant transitions are disabled in case there's a screen tint applied, for some reason. However there was an oversight that caused crossfade/dissolve transition to leave screen filled with a tint colour instead of fading in instantly.

Note that it may be possible to rework transitions and let them work with tint applied, but I'd leave that for the future development.