script interpreter: use modern assembly mnemonics
since the mnemonics aren't used anywhere except debugging, no particular
care has been exerted when designing them (and basically adding them on
an ad-hoc basis), and many of them are ambiguous, outright misleading,
or simply ugly.

for example, there are different instructions for adding a literal to a
register, and adding to 2 registers, yet they have the same mnemonic, making
it harder to reason what happens.
in the new syntax, these now have distinct mnemonics:
addi reg, literal works on an immediate value (thus the i suffix), whereas
add reg1, reg2 works on 2 registers.
the new syntax uses the i suffix consistently for all operations that work
with immediate values.

then there are e.g. "memread", which works on a 4-byte value, whereas
the equivalent mnemonic to work on a 2-byte word is called "memread.w" and
the one for 1-byte access is called "memread.b".
the new syntax calls them memread4, memread2, and memread1, respectively.

the misleading mnemonic "strcmp" was renamed to "streq", since in C strcmp
means a negated value, e.g. if strcmp returns 1, the strings are *not* equal.
other misleading mnemonics such as "checkbounds" and "memwrite.ptr.0" have
also been renamed to something that accurately hints at what the mnemonic
does.

this is the same syntax as used in agsutils assembler and disassembler, making
the 2 finally compatible (the only thing that differs to agsutils now is
"mov", which was left untouched, since it has the awkward semantics of putting
the destination register as second argument, unlike *all* other mnemonics.
agsutils therefore uses a mnemonic called "mr" (move register) instead, which
swaps the dest and source register in the textual form (but it emits the
bytecode in the original order). support for a non-swapping "mov" instruction
can easily be added to agsutils, to have 100% compatibility to AGS.