Engine: refactor timing so all wait loops yield via single function
Based on c4907f5d3d8d432aea03309eb85e62e4f205df59 by sonneveld

This reverts an earlier change made in
6db44217c16054d19f12305257bacb95dc67fd13 that introduced
PollUntilNextFrame that was useful but mixed in the concerns of fast
forward and polling.

Additionally this fixes 100% CPU load in some cases where program is stuck in a loop without yielding.