Script API: changed Object.SetView default loop & frame values to 0
The original Object.SetView function had a somewhat unexpected behavior: if no explicit values were provided for loop and frame then it kept the last loop and frame object had from the previously set view.
This change changes default values to 0, which resets loop and frame if they are not provided.