Editor: reverted to using AudioClip's fixed-index as a reference
This was a big mistake to switch to using AudioClip.ID as a reference to clips in property fields, because currently ID value is volatile and may change if the list of clips was modified in the middle. This commit partially reverts 72a01d64e31a93be900d6e04bd22e812c64608bc and puts fixed index into use again.

This affects ViewFrame's Sound property and GeneralSettings.PlaySoundOnScore.