Engine fonts: new auto outline supports AA TTFs in 32- and 16-bit games
Anti-aliased TTFs require to be alpha-blended, not blit, or the alpha values will be plain copied and final image will be broken.
In 16-bit games we also must use 32-bit stencils, because 16-bit blending can only work correctly if there's an actual color on the destination bitmap (and our intermediate bitmaps are transparent).
8-bit games should not be changed by this update because they don't support font anti-aliasing.