Engine: decoupled maxed fps check from frames_per_second variable
The "frames_per_second" is a game property, that may be set by script. It's not convenient to also use it to store "maxed fps" mode indication. Script may set a different value while the game is in "maxed fps" mode, and this value should be saved normally.

Additionally this supposedly fixes too short delay of a music volume update after speech voice-over, if happened in "maxed fps" mode (see a call to postpone_scheduled_music_update_by() in _displayspeech().