Added test project for AGSEditor
This is one project for the entire AGS editor source code. It may be desirable to create multiple projects, one test project for each project, but I did it like this for convenience.

I also made this using the new csproj format, which may not be compatible with the the currently preferred Visual Studio version. I'll have to rewrite if that is the case

Introduces 4 nuget packages:

- nunit, the actual testing framework
- NUnit3TestAdapter, needed by Visual Studio and later .NET standards to run tests
- NSubstitute, the mocking framework
- Microsoft.NET.Test.Sdk, needed to run the tests via dotnet CLI