Engine: adjusted script opcode mnemonics, partial suggestion of rofl0r
The goal of this change is to make opcode mnemonics more consistent and clear in depicting what opcodes do (also bring closer to what assembler/disassembler from "agsutils" toolset use).
This is a part of rofl0r's suggestion, the other part was left out for now, as it implied more radical changes to these names (may be discussed again later).

The reasoning behind these changes are:

* Remove dots and underscores from the basic arithmetic/logical operations;
* Make every opcode has a distinct mnemonic (remove duplicate mnemonics);
* Arithmetic opcodes that deal with literals have "i" postfix (i - stands for "immediate" term from ASM);
* memread/memwrite mnemonics have number of bytes as a postfix (1,2,4) instead of letter (b for byte, w for word);
* cmp and strcmp mnemonics are replaced with cmpeq/streq, because in C tradition "cmp" means non-binary result, and therefore may be misunderstood.