Fixed NUnit throwing exception
The tests were running fine but the NUnit adapter was throwing an exception on shutdown which could potentially make the CI fail. I don't know exactly why but it had something to do with the path arrangements. Never use real paths in unit tests, save that for integration tests

Had to refactor out the the directories to a interface from AGSEditor class so that they could be mocked