Windows: prevent window from hiding when switching out of fullscreen
We have to call SetWindowPos(HWND_NOTOPMOST) in order to remove WS_EX_TOPMOST style (and let window be a normal kind again), which is a well known solution.
But then a window seem to drop down to the bottom of Z-stack, which is very bad if we are also displaying an error message right after releasing gfx mode.
So, we have to call SetWindowPos(HWND_TOP) in order to bring it back to the top again.