Common: moved String's ptr and length out of the data buffer
Initially String class stored all of its parameters in the externally allocated buffer. This was made mostly for the fastness of copying a string object, hence it was a size of 1 pointer. On downside, however, all operations had to dereference an extra struct pointer first; this included even simpliest reading operations, such as getting string length or a string character (in addition to dereferencing a char ptr itself).

Another serious reason I'd like to change this, however, is to be able to wrap and reference non-owned string buffers, which in turn would allow to wrap string literals without allocating a copy.

After this change String's size will be equal to sizeof(pointer) * 2 + sizeof(size_t).