Engine: explicitly ignore input after service and cutscene controls
Logically, ignoring input mode was designed to prevent misclicks when speech is automatically skipped and player was about to skip it with a click. Therefore, even if this mode is not on, we should still handle service controls (such as windowed/fullscreen switch) mas well as top-level controls (skipping cutscene in this case).

Signed-off-by: Ivan Mogilko <ikm_spb@yahoo.com>