ScriptAPI: implemented Game.IgnoreUserInputForNextTextMs
This property defines a time period, during which if there's a new blocking text on screen, the user input will get disabled until the end of that period.
The purpose of this is to prevent miscliks and occasional speech skipping, and the difference from Game.IgnoreUserInputAfterTextTimeoutMs is that the original property would disable input unconditionally after the text was automatically skipped, while new one does this only if there's another line after first text.