Engine: save current config and adjust setup in RunAGSGame()
Saving config is done in case next game fails to start, so that the last user setup is preserved.

Setup adjustment is a safety hack to let next game initialize with default translation always, because currently there is no way to change translation for game B from game A, and game B won't be able to keep using current game A's translation.