Direct3D: another attempt to keep error message on top after fullscreen
Previous solution that apparently is not reliable: 04c3eb8e49e94254bff4ad15ca22734f36f52677

The problem here is this:
After switching out of the fullscreen mode we have to call SetWindowPos(HWND_NOTOPMOST) in order to remove WS_EX_TOPMOST style (and let window be a normal kind again), which is a well known solution for WinAPI.
But then a window seem to drop down to the bottom of Z-stack, which is very bad if we are also displaying an error message right after releasing gfx mode.
I tried pushing it back by passing HWND_TOP flag, but it's apparent that it does not work very well.

Experiments show that putting application thread to sleep for around 1500 ms would resolve the issue without any extra code. This hints that it may depend on how fast Direct3D switches modes or system restores a window.
But doing so is inconvenient without changing program structure and some routines...

So, for now, I slightly changed they way the window parameters are applied. Now window style is not restored right after the release of current Direct3D mode, but instead a new appropriate style is applied whenever a new mode is set.
Because of this an error message displayed after fullscreen mode will still be shown with TOPMOST flag, but that's a lesser evil. (And this is how it worked in older versions of AGS.)