Editor: fixed Change ID does not alter item location in the flat array
This was an oversight, which made items remain on same positions in the flat list when user changes their IDs, in turn causing inconsistency between ID property and their index in game arrays.

I had to invent somewhat hackish solution to fix this, because of the ugly FolderListHybrid class used to store items in the editor project.