Engine: don't error on missing speech frames if char is not visible
This is mostly a compatibility fix, it seems like older engines did not make this hard test and let things through, and if the character was invisible during speech that did not cause game to crash anyway, as it was not animated nor drawn.
Fixes ending cutscene of the "Trilby's Notes".