Disabled allegro system drivers and added SDL events polling
Unfortunately, game processing is not fit for a single event poll, as game thread may be looping in a certain state for a while before returning to the main loop (such as running a dialog script, waiting for a blocking action to complete, etc). Fixing this is a task for another time.
At the moment we are adding event polls into each sub-state.