SDL-based keyboard input
Based on the work of @sonneveld (Nick Sonneveld) with certain simplifications.

This keeps original AGS keycodes (based on allegro scancodes, although that's irrelevant now) throughout the engine, and only replaces the way it's generated from backend events.
Note that this won't be compatible with unicode text events, and simple AGS code will have to be replaced with something to optionally store unicode text as well.
Also, in the future it may be beneficial to refactor event handling globally, and pass all wanted events to the engine components, either directly, or converted to convenient custom data.