Engine: match game GUIDs when restoring a save, whenever possible
Old logic was testing game file names, probably in case this is a games collection in one directory and save could've done by any of them.
But this test is very unreliable and makes little sense in regular situation. So do GUID test instead, which is far more universal.

What's still in TODO is a second half of the problem: engine should not blindly attempt to load a game using filename from save, but scan for game with the required GUID.