Common: register asset locations with filters, no extra asset searches
AssetManager optionally registers a set of "filters" for the library. It also lets to provide a single filter when searching for an asset. This asset will be only searched in locations that reference this filter.
This is currently used as a way to narrow down search for audio and voice assets that are restricted to particular pakfiles and auxiliary directories.

Had to do this way with filters, because AGS does not currently support a proper asset tree structure, so it's too complicated to e.g. use asset paths like "audio/sound.ext".

Hardcoded filters are: none, "audio" and "voice".

Removed SetLibrarySearch() which toggled asset locations on and off; this proved to be not the best idea.

Now all pakfiles and auxiliary directories are registered in AssetManager with respective filters.
Removed explicit asset searches in directories outside of AssetManager.