Engine: search for config with game path in all common locations
AGS supports config setting that directs to game data location. But it was mistakenly only searched in the "current working dir".
Now rewrote search algorithm: when scanning common locations it first looks for config with instructions, and only if such config is not found it looks for any game data in same place.