Performance fix: Load room images with a image cache
Loading images from unmanaged memory into managed turns out to be very expensive, and it will probably be even more expensive when we move the images to disk. So introducing caching for the room images seems like a good idea.

Before loading images could take a whopping 1400ms on large images like 2784x2816, but now it's reduced down to 0ms.