game data parser: bugfix for versions 2.6-2.7.1
according to analysis of the game data files produced by ags versions
2.6.2, 2.7.0, 2.7.1 and 2.7.2, inventory item names were only added
as of 2.7.0, and dialog script names as of 2.7.2.
the old code used to read past the bounds of the original ac2game.dta
in those versions, effectively reading junk from whatever was stored
behind the data file in the "pack".

closes #1184