Engine: hotfix crashes from setting room object graphic or view >32k
For historical reasons RoomObject stores sprite and view indexes as shorts (int16). But these are in32 in the game data and script api.
It's not clear whether any game has a view ID > 32k, but since there was a raise of sprite limit to 90k in AGS 3.5.0, it's theoretically possible to have a over 32k sprite ID assign to a room object. While this very fix still won't let such ID assigned properly, it will at least prevent from crashes by clamping sprite index to 0 when it overflows int16. Same for the view ID.