Engine: interpret RoomObject's "num" (sprite) field as uint16_t
Sprites are identified as 32-bit integer, however RoomObject struct historically had several fields as shorts (16-bit).
We cannot modify this field's type without breaking plugin compatibility, so for the sake of legacy compatibility have to keep it 2-byte. We may try interpreting it as an unsigned number though, doubling supported range of IDs.