Engine: interpret RoomObject's "view/loop/frame" fields as uint16_t
These are identified as 32-bit integers in game data and script, however RoomObject struct historically had several fields as shorts (16-bit).
We cannot modify this field's type without breaking plugin compatibility, so for the sake of legacy compatibility have to keep it 2-byte. We may try interpreting it as an unsigned number though, doubling supported range of IDs.