Linux: package newest versions of Xiph libraries
Adds build of libogg, libtheora, and libvorbis into the Linux build
Docker container, with installation into /opt. Installing here means
that the existing Debian hook script still gives the same result, but
newer versions are available to be copied into the archive which is
distributed as the Linux version of the engine.

The CI task which builds intermediate archives for each architecture
will now look to replace the named Xiph libraries, if present.
i.e. lib64/libvorbis.so.0 will be replaced by dereferencing
/opt/lib/libvorbis.so.0 when lib64/libvorbis.so.0 exists

This should be seen as a short term change to work around issues found
within older Debian libraries, ideally final archive archive contents
need to be verified to be in the correct state but this needs further
improvements to the build system.