Linux: display launch wrapper name without dereferencing
Given that the primary reason of having the launch wrapper symlinked
would be to provide an interface to one or more launch scripts under an
alternate name or path, it would seem logical that when the output
provides a name it should be what the user specified.

i.e.

$ ./linked_launch --help
Usage: linked_launch [<ags options>]