Engine: Implemented IAGSEngine::GetRenderStageDesc for Plugin API
Historically engine gave much support for plugins which perform software drawing. Ones that perform 3D rendering had to rely on lots of assumptions about how engine sets things up because that information was never fully exposed.
This small amendment to plugin API aims to cover that up at least partially by providing 3 standard transformation matrixes for the current rendering stage.
Because these matrixes are common part of a 3D render, the new interface is not bound to particular renderer class or system. Implementation-wise they are filled in a way compatible with current engine's renderer, so that if plugin is compatible with that renderer then it won't have to convert them.