Engine: implemented "localuserconf" option in cmdline and global config
This options tell the engine to assign user config location to the game directory, unless it's already assigned there by the platform driver.
There's a new cmdline arg "--localuserconf", and a corresponding option in config file.
Because obviously engine needs to use this option before looking up for user file, this option is only read from default game's or global AGS config.

When using this option engine first tests if game dir is writeable. This is done because user config is supposed to be in a writeable location. If it cannot be written to, engine falls back to standard config's location as defined by platform.