Engine: improved the way view clips are remapped for old audio system
Legacy sound system has sounds referenced by an arbitrary number, while new system uses audio clip ID.
For backward compatibility we remap first into second. The old method was replacing original sound number with an audio clip ID in the same field (using a bit flag to indicate its meaning), which also made it be written to saves. This unfortunately made this field reliant on the order audio clips are registered by the engine when scanning the game dir & packs (as legacy games did not specify actual audio clip indexes).

The new method saves remapped audio clip id in the separate field, which is not written to saves. This allows to correctly map legacy sound number to clip even if that save is restored in enviroment with different order or number of clips.