Common: fixed how decrypt_text's result is null-terminated
This seem to be an ancient bug that null terminator is set *before* calling decrypt_text. The thing is that encrypted text also has terminator encrypted, and it cannot be known beforehand what character value it has. Setting null terminator before decryption actually had a chance to break decryption and cause writing outside of the buffer.