Common: picked out data extension read/write algorithm
The algorithm of reading and writing blocks of data is been shared between 3 files now: main game files (where it serves as extensions), room file and translation file. There are minimal differences between them, for example room file has block's numberic ID as a byte and translation file has it as an int32. But it's possible to merge all three in same functions which accept format flags, accomodating for these differences.