Engine: supports reading scripts as separate assets
This allows engine to support game distribs where scripts are not embedded into the main game data, but are present as separate asset files.
First of all, the engine will now treat game.load_compiled_script flag as an indication whether the scripts are found inside main game data file or not.
After the main game data is loaded the engine will look up for script files of hardcoded names for Global and Dialog scripts, and then for the script modules. These are optional. If they are not found no error is raised. If they exist then they will be loaded, replacing any scripts found in the main game data if these were also present.

There are two hardcoded script file names: "GlobalScript.o" and "DialogScript.o".
For the rest of the script modules, engine first tries loading an asset called "ScriptModules.lst" which is supposed to contain simple textual list of filenames. These files are then loaded as script modules in the given order.

Regardless, GlobalScript must exist anyway. If it's not found at all, the error is raised after data load is completed.