ci_find_file: fix omission to check lstat return value, improve fastpath
if lstat() doesn't return success, it's undefined behaviour to access
the uninitialized memory of the stat buffer.

fast path was improved to kick in in almost 100% of cases, namely:
when directory is null and filename starts with a slash (absolute path)
use the absolute path for lookup, when directory is null and filename
doesn't start with a slash (relative path), use current directory.