Common: reorganized ReadExtData functions into a DataExtReader class
This is perhaps how this should've been done from the beginning. ReadExtData is just one function with the loop which takes a "reader callback" function pointer as an argument, but it's not a flexible solution, and makes it difficult to further customize behavior.

This replaces above function with a DataExtParser class which opens blocks and asserts stream position after one has been read, and DataExtReader that is a helper using DataExtParser to read full list of blocks.
DataExtReader should be inherited by subclasses that provide actual data reading implementation, and other overrides when necessary.