Common: optional block over-read leeway, for TRABlockReader
Added DataExtParser::GetOverLeeway() virtual method that lets descendants to justify an acceptable range of mistake for the block length. This may be used to account for malformed files and mistakes in format.

This method is utilized by the TRABlockReader to let load translation files generated by the pre-3.* editors. It appears that the "GameID" block in these reports 1 byte less length than it really has (supposedly by not calculating "game name" field's null terminator.