Engine: restored legacy behavior of MoveCharacterBlocking()
As suggested by @dreammaster, "Maniac Mansion Mania 8" was broken because the game script checks for "MoveCharacterBlocking" return value at some point.
It's probable that this function has not intended to return a value at all (being `void`), and its return value was a remaining value from a program stack. Here we emulate this by returning an actual value.