Common: separated path functions depending on allegro lib
This is done mostly for the stand-alone tools.
It is not easy to link individual parts of allegro lib, because of how its internal dependencies work. Anything depending on allegro cannot be linked without also adding good chunk of the of allegro library.
I think we might later move path/file functionality completely to our engine code.