Filewatcher overhaul
Refactored the FileWathcher utility helper so that each instance maintains its own instance instead of subscribing to a global filewatcher.

The problem with the global filewatcher was that it couldn't be easily disabled/enabled when a file was being saved from the editor. As a result the file watching callback relied on flags and timing to decide if the callback should be invoked or not. This is problematic however since a slow CPU will not be caught by the timing. If we adjust the timing period a fast CPU might not invoke the callback when it's supposed to.

The best solution is to disable the filewatcher when saving internally from the editor