AlFont: apply the face_ascender hack only under the specific condition
It was noted that these hacks fix fonts in some games, but cause regression in others.
After some experiments I found what seems to be a decisive factor here: whether the font's real height is smaller than the requested pixel height.

This change is based on testing several games with different TTF fonts. I am not 100% sure if it's correct, but do this as it produced expected results so far.