AlFont: add a function for getting the true graphical height of a font
Our version of alfont code has certain compatibility hacks that make it simulate the old engine's TTF fonts behavior. Unfortunately this causes the font height returned to not be correct at all times; most commonly it has font's "descent" subtracted.
This may cause the drawn text to be cut off at the bottom, because bitmap surface is not enough to accommodate full height of the glyphs.

This is why we need an extra function to return a real height of the drawn font.