Common: use real font height in game calculations, compat linespacing
This makes a real font height to be used in all game calculations, which would fix some fonts cut off because the prepared bitmap/texture is not large enough to accommodate them.

In sake of the backward compatibility, the fonts which real height is larger than the formal one will keep old linespacing; this is done by setting an explicit linespacing based on the formal font height (except the cases where linespacing is explicitly set in game data).