Engine: tidy SOUNDCLIP's fields (volume, panning) and methods
SOUNDCLIP is a bit complicated as it lets to set volume in both percents and units of 255 range for backward compatibility.
This commit tidies the field names as well as names for setters/getters to make it easier to read what they do.

Removed "panning" field in units of 255, as this is now unnecessary, and having one panning field in percents is enough. (The 255-ranged panning was mostly due to the old sound backend, and not a part of SOUNDCLIP interface).