Engine: removed AudioChannelsLock as it's no longer needed
AudioChannelsLock was added in 3.5.0 as a way to protect SOUNDCLIP objects with a mutex, because SOUNDCLIP array was accessed both on the main game's and audio threads.
In 3.6.0 there's now a separate audio subsystem that does not access SOUNDCLIP objects at all. Instead SOUNDCLIPs have a handle used to issue commands to audio subsystem through its interface.
Thus protecting SOUNDCLIPs is no longer necessary.