Engine: "transactional" SOUNDCLIP
Since having an audio thread, there has been a problem: whenever you work with a playing clip in script, or even adjust its properties as a part of a game update, there's a real chance that these changes will not be all applied to the same sound position. Furthermore, it's possible that the clip will stop playing while you are working with it, preventing you from finishing what you do and causing unexpected and hard to explain behavior.

Here are two examples:

1. You start the sound in script by calling AudioClip.Play, then adjust its playback instance's volume within same script function. There's a chance that first note will play with the default volume, and your changes will only be applied after.

2. You are checking whether there's a clip instance playing on a channel, and if it is - read its properties. E.g.:

    AudioChannel* ch = System.AudioChannel[1];
     if (ch.PlayingClip != null) {
          Label.Text = String.Format("pos: %d", ch.PlayingClip.PositionMs);
     }

There's a small but real chance that the clip stops playing between condition and label assignment, and you will unexpectedly receive a "null pointer" error.

This commit addresses this problem and changes the way SOUNDCLIP works, to ensure following rule:
1) Audio playback's parameters should be only adjusted once per game frame, and not when the user script is running;
2) The engine API must return same clip values (such as Position) throughout one user script pass, these may only change once the control is returned back to the engine and game updates.

This is achieved by making SOUNDCLIP work in "transactions", where it would not apply (most) changes immediately, but save them and sync with the audio subsystem once per a game update.

* SOUNDCLIP does not immediately order sound changes in the audio subsystem; instead it saves them and applies all at once when SOUNDCLIP::update() method is called, once per game frame.
* the playback does not start (or resume) until next update.
* the playback is paused or stopped immediately as told to however.
* SOUNDCLIP reads and saves current clip position only during this update, once per game frame. It reports the saved position when the game script queries it.
* One exception is "seek" function, it is performed immediately and the position updates too. This is to not confuse users with Position property not changing after they called Seek().