Common: in AssetManager separate asset presence check and stream open
Removed AssetLocation struct, and do not fill in any additional data when opening an asset stream. This was useful temporarily in the past when an external code have been using this information to open assets in a special way, but this is no longer the case. All has to be done inside AssetManager, and it's easier to open a stream right away.

Made DoesAssetExist() to use its own code not shared with the OpenAsset routine. This makes slightly more code, but there will be less unnecessary operations with the filesystem in both of them.