Dead end tracking
Track whether a point in code is guaranteed to be unreachable because 'return', 'continue', or 'break' precedes it.
Warn when statements are detected where it is guaranteed that code execution can't pass.

Switch: 'fallthrough;' statement declares that a switch case falls through on purpose.
Switch: Generate a dummy Chunk for the 'default:' case so that the number of chunks corresponds to the number of cases
Switch: Track the code location of the 'default' case with the other cases
Switch: Track the index of the default case separately
Switch: Don't guard the start of the jump table with a jump to end-of-construct if the end of the last case is guaranteed to be unreachable

Nesting stack: Track whether a 'return' etc. has been encountered that jumps out of the current nesting.

Function 'IsSet()' returns whether a BackwardJumpDest has been set.

At end of function body, only generate a '0' return when this point isn't guaranteed to be unreachable.
At end of function body, if the function doesn't return an integer type, warn the user if this point in code can be reached
In return statements, generate a 'RET' opcode instead of a jump to the end of function