Common: apply TTF ascender hack only by the engine's request
Finally seem to found the actual condition when the ascender fix was *NOT* required by the old games. This is when the TTF fonts are drawn with anti-aliasing. The reason for this is uncertain, but the tests done with pre-3.4.1 Windows engine show that the TTF fonts do not have the ascender fix if they are rendered with AA function, and have it if they are rendered with a regular function.