Engine: create a room placeholder to avoid crashes in game_start script
At the time of "game_start" callback no room has been loaded yet. If script tries to access room data at that point, the result is undefined. There are still few functions that access room background or masks, which may cause crashes as these are not yet allocated. On one hand we may add safety checks everywhere, but it may be easy to miss a case; so the most secure solution for now is to create a dummy "room" object with 1x1 8-bit "background" and masks.