Editor: SpeechComponent builds multiple voxes per each Speech subfolder
VOX assets are registered under name equal to the relative filename inside a Speech folder, e.g.:
    Speech/cEgo1.ogg => cEgo1.ogg
    Speech/French/cEgo1.ogg => French/cEgo1.ogg
and so on.

The vox building is now done using DataFileWriter.MakeDataFile() instead of the older function in AGS.Native, because it supports more customization at this point (and it's better to do everything with one method).