Engine: don't mkdirs in the platform driver, return FSLocation instead
Platform drivers used to precreate common directories for the engine and the game; but since we support custom game file locations this is no longer acceptable to unconditionally make these. The platform drivers should only construct the paths to these standard locations, letting the engine decide when to make directories (only if there's a file write operation to the particular place).
These paths are returned as FSLocation struct, which distincts "base" location which assumed to exist, and the sub-directories that engine is allowed to create if they do not exist.