Common: in SpriteFile split WriteRawData() and WriteSpriteData()
There's a separate case when we read the data from the input stream only to write it into another output stream. For that case we want to parse/convert data as less as possible and deal with pure byte arrays. It's not convenient to have same Write* function that both handles writing a byte array and writing a pixel data which may be of variable bitness.