Common: for 3.5.0 games GUI is not interactable if Transparency is 100%
Not an initiall GUI behavior in AGS, but was added by 4fa3ef52163d41aeb8bf3c4573d6b971f17be4c0, which could be unintentional side-effect of that change.
Later I disabled this in ce3e8355e5e20b59a6fa73546669cb25b3084a2f, but sadly there appear to be games already made in 3.5.0 that rely on fully transparent GUIs not being interactable.

Note that there is no change for data format between 3.5.0 and 3.5.1, so we cannot rely on that when implementing compatible behavior. The only option remaining is to use "compiled with" version field, checking that it matches 3.5.0.x.