Common: removed guis list param from GUI:ReadGUI and WriteGUI (useless)
ReadGUI and WriteGUI read and write data using a number of global arrays: "guibut" for Buttons, "guilabel" for Labels, and so forth. In this situation, passing "guis" as an argument is practically pointless, as there are multiple other global objects that will be used along. Also, having this parameter may give a wrong impression to programmers that this function does NOT rely on global objects, while it does.

The proper solution would be perhaps to remove global list declarations from the guimain header, and pass ALL the lists into these read/write functions (packed in a struct maybe).