Engine: disabled video framedrop on all platforms.
After porting to SDL2/OpenAL, APEG library logic seem to be no longer fully suitable, and it detect audio ahead of video way too often, which may lead to video dropping all the frames trying to catch up, which never works in practice. I think the good solution might be to rewrite the video playback (or use another library), but this fix should work for the time being.