Engine: OpenAlDecoder reports position with account to playback speed
To be able to do so we have to remember playback speed per queued buffer and apply as necessary.

Unfortunately, there's still a mistake somewhere, which I was not able to track to the end: because of it the reported position may jump slightly backwards when the playback speed changes. Usually this happens when speed goes from high to low. As a dirty fix we remember the last reported position and clamp new report to it.