Editor: removed a rule to not export Get/SetProperty args from script
This was supposed to prevent "unnecessary" export of property names, but with the introduction of the SetProperty functions prevented the export of their strings also. Overall this seem to be a rather dubious attempt at optimizing things. Besides it does not really work correctly, as simply splitting the function call into multiple lines would let to walk around this prevention.