apeg: non-global setjmp buffer, and set jmp points in the new functions
This moves setjmp buffer and error message buffer to APEG_LAYER, to potentially allow multiple simultaneous video playbacks with apeg library in the future.
Also setup jmp points in the new functions created by the previous commit.

I'm not happy that we have to deal with setjmp, longjmp in this library, but removing them would also require to redesign error handling to ensure all functions correctly return back on error.