Engine: reimplement audio playback pos using buffer timestamps
Instead of calculating current position from the size of processed al buffers, save expected timestamps in the SDLDecoder and pass along with the buffer. Use these timestamps to know which position we just passed when processing them in OpenAlSource.
Let OpenAlSource to also predict next timestamp based on input data and use this prediction in case next buffer does not contain a valid timestamp, as a last resort.