Engine: fixed TheoraPlayer incorrectly placing videos of certain ratios
According to the documentation: encoded theora frames must be a multiple of 16 in width and height.
Which means that the original content may end up positioned on a larger frame. In such case we store this surface in a separate wrapper for the reference, while the actual video frame is assigned a sub-bitmap (a portion of the full frame).